ENG-4983 set bundle as removed only for app-engine uninstall success,          added more logs and fix for WidgetDescriptor          filtered components by usage          added fix to use only existing components in core client request for delete enabling subsequents retries          added more test          fixed integration test wrong expected value          added optimization on error management

+++ src/main/java/org/entando/kubernetes/service/digitalexchange/job/EntandoBundleUninstallService.java
@@ -23,6 +23,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
 import org.entando.kubernetes.client.model.EntandoCoreComponentDeleteRequest;
 import org.entando.kubernetes.client.model.EntandoCoreComponentDeleteResponse;
+import org.entando.kubernetes.client.model.EntandoCoreComponentDeleteResponse.EntandoCoreComponentDelete;
 import org.entando.kubernetes.client.model.EntandoCoreComponentDeleteResponse.EntandoCoreComponentDeleteStatus;
 import org.entando.kubernetes.exception.EntandoComponentManagerException;
 import org.entando.kubernetes.exception.digitalexchange.BundleNotInstalledException;
@@ -31,6 +32,7 @@ import org.entando.kubernetes.exception.job.JobConflictException;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.installable.Installable;
 import org.entando.kubernetes.model.bundle.processor.ComponentProcessor;
+import org.entando.kubernetes.model.bundle.usage.ComponentUsage;
 import org.entando.kubernetes.model.job.EntandoBundleComponentJobEntity;
 import org.entando.kubernetes.model.job.EntandoBundleEntity;
 import org.entando.kubernetes.model.job.EntandoBundleJobEntity;
@@ -59,6 +61,8 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
     private final @NonNull EntandoCoreClient entandoCoreClient;
     private final @NonNull ObjectMapper objectMapper;
 
+    private static final String KEY_SEP = "_";
+
     public EntandoBundleJobEntity uninstall(String componentId) {
         EntandoBundleEntity installedBundle = installedComponentRepository.findByBundleCode(componentId)
                 .orElseThrow(() -> new BundleNotInstalledException("Bundle " + componentId + " is not installed"));
@@ -67,7 +71,9 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
 
         verifyBundleUninstallIsPossibleOrThrow(installedBundle);
 
-        return createAndSubmitUninstallJob(installedBundle.getJob());
+        List<ComponentUsage> componentUsages = verifyAndGetListOfComponentUsageOrThrow(installedBundle);
+
+        return createAndSubmitUninstallJob(installedBundle.getJob(), componentUsages);
     }
 
     private void verifyBundleUninstallIsAllowedOrThrow(String bundleCode) {
@@ -81,7 +87,6 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
 
     private void verifyBundleUninstallIsPossibleOrThrow(EntandoBundleEntity bundle) {
         if (bundle.getJob() != null && bundle.getJob().getStatus().equals(JobStatus.INSTALL_COMPLETED)) {
-            // verifyNoComponentInUseOrThrow(bundle);
             verifyNoConcurrentUninstallOrThrow(bundle);
         } else {
             throw new EntandoComponentManagerException(
@@ -100,16 +105,22 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
         }
     }
 
-    private void verifyNoComponentInUseOrThrow(EntandoBundleEntity bundle) {
+    private List<ComponentUsage> verifyAndGetListOfComponentUsageOrThrow(EntandoBundleEntity bundle) {
         List<EntandoBundleComponentJobEntity> bundleComponentJobs = compJobRepo.findAllByParentJob(bundle.getJob());
-        if (bundleComponentJobs.stream()
-                .anyMatch(e -> usageService.getUsage(e.getComponentType(), e.getComponentId()).getUsage() > 0)) {
-            throw new JobConflictException(
-                    "Some of bundle " + bundle.getId() + " components are in use and bundle can't be uninstalled");
-        }
+        return usageService.getComponentsUsageDetails(bundleComponentJobs).stream()
+                .filter(componentUsage -> {
+                    if (componentUsage.getHasExternal()) {
+                        throw new JobConflictException(String.format(
+                                "Component '%s' of bundle '%s' contains external references and bundle can't be "
+                                        + "uninstalled", componentUsage.getCode(), bundle.getId()));
+                    }
+                    return true;
+                }).collect(Collectors.toList());
     }
 
-    private EntandoBundleJobEntity createAndSubmitUninstallJob(EntandoBundleJobEntity lastAvailableJob) {
+
+    private EntandoBundleJobEntity createAndSubmitUninstallJob(EntandoBundleJobEntity lastAvailableJob,
+            List<ComponentUsage> componentUsages) {
 
         EntandoBundleJobEntity uninstallJob = new EntandoBundleJobEntity();
         uninstallJob.setComponentId(lastAvailableJob.getComponentId());
@@ -119,12 +130,13 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
         uninstallJob.setProgress(0.0);
         EntandoBundleJobEntity savedJob = jobRepo.save(uninstallJob);
 
-        submitUninstallAsync(uninstallJob, lastAvailableJob);
+        submitUninstallAsync(uninstallJob, lastAvailableJob, componentUsages);
 
         return savedJob;
     }
 
-    private void submitUninstallAsync(EntandoBundleJobEntity parentJob, EntandoBundleJobEntity referenceJob) {
+    private void submitUninstallAsync(EntandoBundleJobEntity parentJob, EntandoBundleJobEntity referenceJob,
+            List<ComponentUsage> componentUsages) {
         CompletableFuture.runAsync(() -> {
             JobTracker<EntandoBundleJobEntity> parentJobTracker = new JobTracker<>(parentJob, jobRepo);
             JobScheduler scheduler = new JobScheduler();
@@ -162,25 +174,41 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
                     optCompJob = scheduler.extractFromQueue();
                 }
 
-                JobStatus finalStatus = executeDeleteFromAppEngine(componentToUninstallFromAppEngine, parentJob);
+                parentJobResult.clearException();
+                JobStatus finalStatus = executeDeleteFromAppEngine(
+                        keepNeededComponentsOnly(componentUsages, componentToUninstallFromAppEngine),
+                        parentJobResult);
                 parentJobResult.setStatus(finalStatus);
 
-                installedComponentRepository.deleteByBundleCode(parentJob.getComponentId());
+                if (JobStatus.UNINSTALL_COMPLETED.equals(finalStatus)) {
+                    installedComponentRepository.deleteByBundleCode(parentJob.getComponentId());
+                } else {
+                    parentJobResult.setUninstallException(new EntandoComponentManagerException(parentJob.getUninstallErrors()));
+                }
 
                 uninstallProgress.increment();
-                parentJobResult.clearException();
                 parentJobResult.setProgress(1.0);
 
-                log.info("Component '{}' uninstalled successfully", parentJob.getComponentId());
+                log.info("Uninstall operation completed with status '{}' for component:'{}'",
+                        finalStatus,
+                        parentJob.getComponentId());
 
             } catch (Exception ex) {
                 log.error("An error occurred while uninstalling component " + parentJob.getComponentId(), ex);
                 parentJobResult.setStatus(JobStatus.UNINSTALL_ERROR);
-                parentJobResult.setInstallException(ex);
+                parentJobResult.setUninstallException(ex);
             }
 
             parentJobTracker.finishTracking(parentJobResult);
         });
+    }
+
+    private List<EntandoBundleComponentJobEntity> keepNeededComponentsOnly(List<ComponentUsage> componentUsages,
+            List<EntandoBundleComponentJobEntity> componentToUninstallFromAppEngine) {
+        return componentToUninstallFromAppEngine
+                .stream()
+                .filter(cje -> filterComponentToDelete(cje, componentUsages))
+                .collect(Collectors.toList());
     }
 
     private JobTracker<EntandoBundleComponentJobEntity> trackExecution(EntandoBundleComponentJobEntity cj,
@@ -194,6 +222,7 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
 
     private Queue<EntandoBundleComponentJobEntity> createUninstallComponentJobs(EntandoBundleJobEntity parentJob,
             EntandoBundleJobEntity referenceJob) {
+
         List<EntandoBundleComponentJobEntity> installJobs = compJobRepo.findAllByParentJob(referenceJob);
         return installJobs.stream()
                 .map(cj -> {
@@ -209,6 +238,16 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
                         .comparingInt(cj -> ((EntandoBundleComponentJobEntity) cj).getInstallable().getPriority())
                         .reversed())
                 .collect(Collectors.toCollection(ArrayDeque::new));
+    }
+
+    private boolean filterComponentToDelete(EntandoBundleComponentJobEntity cje,
+            List<ComponentUsage> componentUsages) {
+        return componentUsages.stream()
+                .filter(componentUsage -> componentUsage.getType().equals(cje.getComponentType().getTypeName())
+                        && componentUsage.getCode().equals(cje.getComponentId()))
+                .findFirst()
+                .map(ComponentUsage::isExist)
+                .orElse(true);
     }
 
     private JobResult executeUninstall(Installable<?> installable) {
@@ -231,7 +270,7 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
     }
 
     private JobStatus executeDeleteFromAppEngine(List<EntandoBundleComponentJobEntity> toDelete,
-            EntandoBundleJobEntity parentJob) {
+            JobResult parentJobResult) {
         EntandoCoreComponentDeleteResponse response = entandoCoreClient.deleteComponents(toDelete.stream()
                 .map(EntandoCoreComponentDeleteRequest::fromEntity)
                 .flatMap(Optional::stream)
@@ -240,11 +279,13 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
         switch (response.getStatus()) {
             case FAILURE:
                 log.debug("All deletes are in error with response:'{}'", response);
-                markError(parentJob, response);
+                markGlobalError(parentJobResult, response);
+                markSingleErrors(toDelete,response);
                 return JobStatus.UNINSTALL_ERROR;
             case PARTIAL_SUCCESS:
                 log.debug("Partial deletes are in error with response:'{}'", response);
-                markError(parentJob, response);
+                markGlobalError(parentJobResult, response);
+                markSingleErrors(toDelete,response);
                 return JobStatus.UNINSTALL_PARTIAL_COMPLETED;
             case SUCCESS:
             default:
@@ -253,14 +294,61 @@ public class EntandoBundleUninstallService implements EntandoBundleJobExecutor {
         }
     }
 
-    private void markError(EntandoBundleJobEntity parentJob, EntandoCoreComponentDeleteResponse response) {
+    private void markSingleErrors(List<EntandoBundleComponentJobEntity> toDelete,
+            EntandoCoreComponentDeleteResponse response) {
+        // convenience map to speed up the search of an EntandoBundleComponentJobEntity related to a specific component
+        // included in the app-engine response
+        Map<String, EntandoBundleComponentJobEntity> toDeleteMap = toDelete.stream()
+                .collect(Collectors.toMap(this::composeComponentJobEntityUniqueKey,Function.identity()));
+
+        response.getComponents()
+                .stream()
+                .filter(c -> EntandoCoreComponentDeleteStatus.FAILURE.equals(c.getStatus()))
+                .map(entandoCoreComponentDelete ->
+                        // find the EntandoBundleComponentJobEntity, related to the ith component of the response,
+                        // to set on it the appropriate error message
+                        Optional.ofNullable(
+                                        toDeleteMap.get(composeComponentDeleteUniqueKey(entandoCoreComponentDelete)))
+                                .orElseThrow(() -> new IllegalStateException(
+                                        String.format("Missing job for component '%s' of type '%s'",
+                                                entandoCoreComponentDelete.getCode(),
+                                                entandoCoreComponentDelete.getType())))
+                ).forEach(cje -> {
+                    // set the error message and code and save to DB
+                    cje.setUninstallErrorMessage("Error in deleting component from app-engine");
+                    cje.setUninstallErrorCode(100);
+                    compJobRepo.save(cje);
+                });
+    }
+
+    private String composeComponentDeleteUniqueKey(EntandoCoreComponentDelete entandoCoreComponentDelete) {
+        if (entandoCoreComponentDelete == null
+                || StringUtils.isEmpty(entandoCoreComponentDelete.getCode())
+                || StringUtils.isEmpty(entandoCoreComponentDelete.getType())) {
+            throw new IllegalArgumentException("Error in composing key from Entando Core Component Delete: "
+                    + "element, code or type fields have null or empty values");
+        }
+        return entandoCoreComponentDelete.getCode() + KEY_SEP + entandoCoreComponentDelete.getType();
+    }
+
+    private String composeComponentJobEntityUniqueKey(EntandoBundleComponentJobEntity componentJobEntity) {
+        if (componentJobEntity == null
+                || StringUtils.isEmpty(componentJobEntity.getComponentId())
+                || StringUtils.isEmpty(componentJobEntity.getComponentType().getTypeName())) {
+            throw new IllegalArgumentException("Error in composing key from Entando Bundle Component Job Entity: "
+                    + "element, code or type fields have null or empty values");
+        }
+        return componentJobEntity.getComponentId() + KEY_SEP + componentJobEntity.getComponentType().getTypeName();
+    }
+
+    private void markGlobalError(JobResult parentJobResult, EntandoCoreComponentDeleteResponse response) {
         try {
             String uninstallErrors = objectMapper.writeValueAsString(response.getComponents().stream()
                     .filter(c -> !EntandoCoreComponentDeleteStatus.SUCCESS.equals(c.getStatus())).collect(
                             Collectors.toList()));
             String valueToSave = StringUtils.truncate(uninstallErrors, MAX_COMMON_SIZE_OF_STRINGS);
             log.debug("save error inside parentJob:'{}'", valueToSave);
-            parentJob.setUninstallErrors(valueToSave);
+            parentJobResult.setUninstallException(new EntandoComponentManagerException(valueToSave));
 
         } catch (JsonProcessingException ex) {
             log.error("with response:'{}' we had a json error", response, ex);
